requirements.txt
pyautogui
pywin32
pyyaml
